Refactor database dump and load functions
Refactoring occurred on dumpDatabaseToJson.ts file mainly to extract the logic of saving data into a separate function named saveJsonToDump. Also, the function loadJsonToDatabase that was previously in dumpDatabaseToJson.ts has been moved to a new file called loadDatabaseFromJson.ts. The refactoring aimed at improving code maintainability and readability.
